Rethink how dismissal IDs are handled
The only time a notice really needs an ID is if we're planning on making it dismissible **and** we want to be able to track it.

Furthermore, we should only be tracking dismissed notices that have an ID associated with them — we don't care that someone hid the "Settings updated successfully!"-type messages.

To this end, the `$id` property has been renamed to `$dismissibileKey`, and is only set via the `setDismissible()` method.

For flexibility, this may operate in a few different ways:

```php
// Do not make the notice dismissible (default behavior)
$notice->setDismissible(false);

// Enable the notice to be dismissed, but do not track dismissals
$notice->setDismissible(true);

// Track dismissals of this notice with key "some-key"
$notice->setDismissible(true, 'some-key');

// Track dismissals of this notice, but generate a key
$notice->setDismissible(true, true);
```

Generated keys take the form of `{type}:{hash}`, where hash is the first 10 characters of an MD5 checksum of the `$message` property.